requirements.txt
numpy
opencv-python-headless
pillow
tensorflow
streamlit